* Change dF / F0 calculation from Sun & Sudhof 12;5% of trace and bottom
10% values to MAD estimation and peak masking

* Update peak detection functions to mask dF / F0 using MAD 

* Threshold stays at 4.5 * SD (SD estimated from MAD calculation)

* Fix config_loader functions to load default and user made config.json
files

* Fix run_suite2p code to work with either default or user made
config.json files --> Improves pipeline functionality

* Correct previous mistake in update_iscell and synapse v dendrite
classifications using getStats in csv generation

* Create Docstrings for all functions in all files in line with PEP 8
and PEP 257

+++ src/analyze_suite2p/detector_utility.py
@@ -6,13 +6,32 @@ import scipy.signal as signal
 from scipy.stats import norm
 from analyze_suite2p import config_loader
 from BaselineRemoval import BaselineRemoval
+import os
 
-config = config_loader.load_json_config_file()
+_DEFAULT_CONFIG = config_loader.load_json_config_file()
+config = _DEFAULT_CONFIG
 
-def calculate_deltaF(F_file):
-    """Function to calculated dF from F and Fneu of suite2p based on Sun & Sudhof, 2019 dF/F calculations
-    inputs: 
-    F_file: F.npy file that serves as a template for understanding the fluorescence of individual ROIs"""
+def calculate_deltaF(F_file, event_threshold = 2):
+    """
+    Convert raw fluorescence (F.npy) into change in fluorescence compared to baseline (dF / F0).
+
+    Args:
+    -----------
+    F_file : str
+        Path to NumPy array containing raw flourescence (F.npy) trace from suite2p.
+    
+    event_threshold: float
+        Threshold (in MAD units) to mask obvious events by multiplying threshold by standard deviation. 
+        The Default value is 3; smaller values will limit the number of baseline points used for correction.
+
+    Returns:
+    --------
+    deltaF : 1D numpy array
+        dF/F0 normalized fluorescence
+        MAD baseline estimated
+        ZhangFit / airPLS automated baseline correction
+        deltaF is saved into the suite2p output folder generated from suite2p ROI detection.
+    """
 
     savepath = rf"{F_file}".replace("\\F.npy","") ## make savepath original folder, indicates where deltaF.npy is saved
     F = np.load(rf"{F_file}", allow_pickle=True)
@@ -20,27 +39,90 @@ def calculate_deltaF(F_file):
     deltaF= []
     for f, fneu in zip(F, Fneu):
         corrected_trace = f - (0.7*fneu) ## neuropil correction
-        amount = int(0.125*len(corrected_trace))
-        middle = 0.5*len(corrected_trace)
-        F_sample = (np.concatenate((corrected_trace[0:amount], corrected_trace[int(middle-amount/2):int(middle+amount/2)], 
-                    corrected_trace[len(corrected_trace)-amount:len(corrected_trace)])))  #dynamically chooses beginning, middle, end 12.5%, changeable
-        #TODO decide if mean, median or mode is best for deltaF calculations
-        F_baseline = np.percentile(F_sample, 10)
-        normalized_F = (corrected_trace-F_baseline)/F_baseline
+        trace_median = np.median(corrected_trace)
+        trace_mad = np.median(np.abs(corrected_trace - trace_median))
+        norm_sigma = 1.4826*trace_mad
+        baseline_mask = np.abs(corrected_trace - trace_median) < event_threshold * norm_sigma
+
+        F0 = np.median(corrected_trace[baseline_mask])
+        # amount = int(0.125*len(corrected_trace))
+        # middle = 0.5*len(corrected_trace)
+        # F_sample = (np.concatenate((corrected_trace[0:amount], corrected_trace[int(middle-amount/2):int(middle+amount/2)], 
+        #             corrected_trace[len(corrected_trace)-amount:len(corrected_trace)])))  #dynamically chooses beginning, middle, end 12.5%, changeable
+        # #TODO decide if mean, median or mode is best for deltaF calculations
+        # F_baseline = np.percentile(F_sample, 10)
+        normalized_F = (corrected_trace-F0)/F0
         baseline_correction = BaselineRemoval(normalized_F)
         ZhangFit_normalized = baseline_correction.ZhangFit(lambda_= 10)
         deltaF.append(ZhangFit_normalized)
         
     deltaF = np.array(deltaF)
     deltaF = np.squeeze(deltaF)
-    np.save(f"{savepath}/deltaF.npy", deltaF, allow_pickle=True)
-    print(f"delta F calculated for {F_file[len(config.general_settings.main_folder)+1:-21]}")
-    print(f"delta F traces saved as deltaF.npy under {savepath}\n")
+    if not os.path.exists(f"{savepath}/deltaF.npy"):
+        np.save(f"{savepath}/deltaF.npy", deltaF, allow_pickle=True)
+        print(f"delta F traces saved as deltaF.npy under {savepath}\n")
+    else:
+        print(f"deltaF files already exist for {F_file[len(config.general_settings.main_folder)+1:-21]}")
+
     return deltaF
 
 
+def estimate_single_trace_baseline_noise_mad(F_trace, event_threshold = 2):
+    """
+    Estimate noise sigma from baseline-only windows using MAD.
+    
+    Args:
+    -----------
+    F : 1D numpy array
+        Baseline-corrected ΔF/F trace.
+    frame_rate : float
+        Sampling rate (Hz).
+    event_threshold: float
+        Preserved from calculate_deltaF function above.
+        Threshold (in MAD units) to mask obvious events by multiplying by estimated noise standard deviation. 
+        Default: 2 (SD above median)
+        Smaller values will limit the number of baseline points used for correction.
+    min_baseline_sec : float
+        Minimum duration (seconds) of a baseline window.
+        Default: 10 s
+
+    Returns:
+    --------
+    sigma : float
+        Estimated noise standard deviation.
+    baseline_mask : boolean array
+        Mask of samples classified as baseline.
+    """
+
+    trace_median = np.median(F_trace)
+    mad = np.median(np.abs(F_trace - trace_median))
+    sigma = 1.4826 * mad
+    event_mask = np.abs(F_trace - trace_median) > event_threshold * sigma
+
+    trace_baseline = ~event_mask
+
+    baseline_samples = F_trace[trace_baseline]
+    
+    baseline_median = np.median(baseline_samples)
+    baseline_mad = np.median(np.abs(baseline_samples - baseline_median))
+    sigma = 1.4826 * baseline_mad
+    
+    return sigma, baseline_samples
 
 def filter_outliers(trace):
+    """
+    Filter outliers (peaks) from calcium trace using the trace IQR
+
+    Args:
+    -----------
+    trace : 1D numpy array
+        Fluorescence trace (e.g., F.npy) from suite2p output.
+    
+    Returns:
+    --------
+    filtered_values : 1D array
+        values from 1D array that fall within the original trace IQR.
+    """
     q1,q3 = np.percentile(trace, [25,75])
     iqr = q3-q1
     lower_bound = q1 - 1.5*iqr
@@ -49,19 +131,51 @@ def filter_outliers(trace):
     return filtered_values   
 
 
-def single_synapse_baseline_correction_and_peak_return(deltaF, return_peaks = False, 
+def single_synapse_peak_detection(deltaF, return_peaks = False, 
                                                        return_decay_frames = False, 
                                                        return_amplitudes = False, 
                                                        return_decay_time = False,
-                                                       return_peak_count = False):
-    """this function takes a single time series data series and converts into deltaF / F; it then will return, frames where peaks occurred, amplitudes of peaks
-    the number of peaks detected, the decay frames (TBD) and the decay time converted into sections"""
-    iqr_noise = filter_outliers(deltaF) #iqr noise
-    mu, std = norm.fit(iqr_noise) #median and sd of noise of trace based on IQR
-    peak_detection_multiplier = float(config.analysis_params.peak_detection_threshold)
-    threshold = mu + (peak_detection_multiplier * std)
-    peaks, _ = find_peaks(deltaF, height = threshold, distance = 5)
-    amplitudes = deltaF[peaks] - mu #amplitude
+                                                       return_peak_count = False, 
+                                                       extract_peaks = False):
+    """
+    Identify time stamps and metrics of individual calcium spikes for a single ROI.
+    Args:
+    -----------
+    deltaF: 1D numpy array
+        Normalized fluroescence trace 
+    return_peaks: bool, optional
+        Returns time stamps (frame) for each peak
+    return_decay_frames: bool, optional
+        Returns time stamp (frame) for when each peak returns to threshold
+        if no return to threshold --> returns NaN
+    return_amplitudes: bool, optional
+        Returns normalized amplitude for each detected peak
+    return_decay_times: bool, optional
+        Returns decay time from peak frame to crossing threshold (in seconds)
+    return_peak_count: bool, optional
+        Returns len(peaks)
+    extract_peaks: bool, optional
+        Returns peaks + 30 frames for peak library --> to be used for Tau calculations
+
+    Returns:
+    --------
+    IF any==True:
+        return_peaks: returns calcium spike time_stamps
+        return_decay_frames: returns number of frames for calcium spike to decay to threshold
+        return_amplitudes: returns amplitude of calcium spike in relation to baseline fluorescence (F0)
+        return_decay_time: returns decay time in seconds (converts number of frames into seconds)
+        return_peak_count: returns the total number of calcium spikes for the ROI fluorescence trace
+        extract_peaks: returns deltaF window around calcium spike for calcium spike library 
+    """
+    
+    sigma, deltaF_baseline = estimate_single_trace_baseline_noise_mad(deltaF, event_threshold=2)
+    
+    baseline_reference = np.median(deltaF_baseline)
+    peak_detection_multiplier = 4.5# float(config.analysis_params.peak_detection_threshold)
+    threshold = np.median(deltaF_baseline) + (peak_detection_multiplier * sigma)
+
+    peaks, _ = find_peaks(deltaF, height = threshold, distance = 5, prominence = baseline_reference + sigma, width = (2,None))
+    amplitudes = deltaF[peaks] - baseline_reference #amplitude
     peak_count = len(peaks)
     negative_points = np.where((deltaF < threshold))[0]
 
@@ -102,11 +216,29 @@ def single_synapse_baseline_correction_and_peak_return(deltaF, return_peaks = Fa
         return decay_time
     if return_peak_count == True:
         return peak_count
-    # if calculate_tau == True:
-    #     return decay_time
+    if extract_peaks:
+        peak_dict = {}
+        #TODO fix up these points for peak library
+        # for peak in peaks:
+        #     peak_dict.update(f'peak_{peak}': 'deltaF[peak:peak+30]')
+        
+        return peak_dict
 
 
 def detect_spikes_by_mod_z(input_trace, **signal_kwargs):
+    """
+    Detect spikes by median absolute difference (MAD) of each frame from median of the trace.
+
+    Args:
+    -----------
+    input_trace: 1D NumPy array
+    **signal_kwargs: assorted see signal.find_peaks()
+        Ex. width = (min,max), peak_prominence = type(float), height = type(float), threshold = type(float), distance = int/float
+    Returns:
+    --------
+    peak time frames using signal.find_peaks() function
+
+    """
     median = np.median(input_trace)
     deviation_from_med = np.array(input_trace) - median
     mad = np.median(np.abs(deviation_from_med))
@@ -115,6 +247,24 @@ def detect_spikes_by_mod_z(input_trace, **signal_kwargs):
 
 
 def plot_spikes(raw_trace, detector_func, detector_trace=None, **detector_kwargs):
+    """
+    Plot ROI calcium trace with overlayed detected spikes as red vertical lines.
+
+    Args:
+    -----------
+    raw_trace: 1D NumPy array
+    detector_func: Function
+        Ex. scipy.signal.find_peaks() / detect_spikes_by_mod_z()
+    detector_trace: bool, optional
+    **detector_kwargs: assorted, optional
+        Ex. scipy.signal.find_peaks(x, height = , threshold = , peak_prominence = , width = , distance = )
+
+    Returns:
+    -------- 
+    matplotlib.pyplot.plot line graph
+        blue: detector_trace (if true) or raw trace
+        red: detected spikes
+    """
     if detector_trace is None:
         detector_input_trace = raw_trace.copy()
     else:
@@ -127,15 +277,46 @@ def plot_spikes(raw_trace, detector_func, detector_trace=None, **detector_kwargs
 
 
 def rolling_min(input_series, window_size):
+    """
+    Calculate rolling minimum value (input_series.rolling()) over different windows of the input trace.
+
+    Args:
+    -----------
+    input_series: 1D NumPy array
+        raw_trace / F.npy / deltaF.npy
+    window_size: int
+        Size of window to measure with each iteration
+
+    Returns:
+    -------- 
+    m: int / float
+        Smallest local minimum across all windows
+    """
     r = input_series.rolling(window_size, min_periods=1)
     m = r.min()
     return m
 
 
 def remove_bleaching(input_trace):
+    """
+    Basic first-order polynomial function to remove bleaching from single ROI calcium imaging trace
+
+    Args:
+    -----------
+    input_trace: 1D array
+        raw fluorescence trace (F.npy or corrected: F.npy - 0.7*Fneu.npy)
+        functions by processing one ROI at a time
+
+    Returns:
+    --------
+    input_trace - fit(range(len(input_trace)))
+        input trace adjusted by rolling minimum
+        polynomial fit built on length of trace, rolling min values, and order of polynomial (e.g. 2nd)
+        poly1d fits a 1 dimensional polynomial to the adjusted trace which is subtraced from the raw trace (input_Trace)
+
+    """
     min_trace = rolling_min(pd.Series(input_trace), window_size=int(len(input_trace)/10))
     fit_coefficients = np.polyfit(range(len(min_trace)), min_trace, 2)
     fit = np.poly1d(fit_coefficients)
     return input_trace - fit(range(len(input_trace)))
-
